+++ kopy-compiler/main/kotlin/com/javiersc/kotlin/kopy/compiler/ir/transformers/IrFunctionsTransformer.kt
@@ -10,6 +10,7 @@ import com.javiersc.kotlin.compiler.extensions.ir.regularParameters
 import com.javiersc.kotlin.kopy.compiler.atomicReferenceClassId
 import com.javiersc.kotlin.kopy.compiler.copyName
 import com.javiersc.kotlin.kopy.compiler.invokeName
+import com.javiersc.kotlin.kopy.compiler.ir.utils.insertExtensionReceiver
 import com.javiersc.kotlin.kopy.compiler.ir.utils.isKopyCopy
 import com.javiersc.kotlin.kopy.compiler.ir.utils.isKopyCopyOrInvoke
 import com.javiersc.kotlin.kopy.compiler.ir.utils.isKopySet
@@ -68,11 +69,13 @@ internal class IrFunctionsTransformer(
     private val mapFunction: IrSimpleFunction
         get() =
             pluginContext
-                .referenceFunctions(mapCallableId)
+                .finderForBuiltins()
+                .findFunctions(mapCallableId)
                 .firstIsInstance<IrSimpleFunctionSymbol>()
                 .owner
 
-    private val listClass: IrClass = pluginContext.referenceClass(listClassId)!!.owner
+    private val listClass: IrClass =
+        pluginContext.finderForBuiltins().findClass(listClassId)!!.owner
 
     override fun visitSimpleFunction(declaration: IrSimpleFunction): IrStatement {
         fun originalFunction(): IrStatement = super.visitSimpleFunction(declaration)
@@ -91,104 +94,105 @@ internal class IrFunctionsTransformer(
 
     context(context: IrPluginContext)
     private fun transformCopyOrInvokeFunction(declaration: IrSimpleFunction): IrSimpleFunction {
-        declaration.body =
-            DeclarationIrBuilder(declaration.symbol).irBlockBody {
-                val thisCopyIrVariable: IrVariable = thisCopyIrVariable(declaration)
-                +copyCall(declaration, thisCopyIrVariable)
-                +atomicValueIrReturn(declaration, thisCopyIrVariable)
-            }
+        declaration.body = declaration.createCopyOrInvokeBody()
         return declaration
     }
 
     context(context: IrPluginContext)
-    private fun transformSetFunction(declaration: IrSimpleFunction): IrSimpleFunction {
-        val function: IrSimpleFunction = declaration.apply {
-            body =
-                DeclarationIrBuilder(declaration.symbol).irBlockBody {
-                    val unitClass: IrClassSymbol = context.irBuiltIns.unitClass.owner.symbol
-                    val unitCall: IrGetObjectValue = irGetObject(unitClass)
-                    val irReturn: IrReturn = irReturn(unitCall)
-                    +irReturn
-                }
+    private fun IrSimpleFunction.createCopyOrInvokeBody() =
+        DeclarationIrBuilder(symbol).irBlockBody {
+            val thisCopyIrVariable: IrVariable =
+                createTemporaryCopyVariable(this@createCopyOrInvokeBody)
+            +callCopyLambda(this@createCopyOrInvokeBody, thisCopyIrVariable)
+            +returnAtomicValue(this@createCopyOrInvokeBody, thisCopyIrVariable)
         }
-        return function
+
+    context(context: IrPluginContext)
+    private fun transformSetFunction(declaration: IrSimpleFunction): IrSimpleFunction {
+        declaration.body = declaration.createReturnUnitBody()
+        return declaration
     }
 
     context(context: IrPluginContext)
-    private fun transformUpdateFunction(declaration: IrSimpleFunction): IrSimpleFunction {
-        val function: IrSimpleFunction = declaration.apply {
-            body =
-                DeclarationIrBuilder(declaration.symbol).irBlockBody {
-                    val thisGet: IrGetValue = irGet(extensionReceiver!!)
-                    val transformValueParameter: IrValueParameter = regularParameters.first()
-                    val func1Invoke: IrSimpleFunction =
-                        function1Class.findDeclaration<IrSimpleFunction> { it.name == invokeName }!!
-
-                    val transformCallDispatchReceiver: IrGetValue =
-                        irGet(transformValueParameter).apply {
-                            origin = IrStatementOrigin.VARIABLE_AS_FUNCTION
-                        }
-
-                    val transformCall: IrMemberAccessExpression<*> =
-                        irCallOp(
-                            callee = func1Invoke.symbol,
-                            type = declaration.typeParameters.first().defaultType,
-                            dispatchReceiver = transformCallDispatchReceiver,
-                            argument = thisGet,
-                            origin = IrStatementOrigin.INVOKE,
-                        )
-                    +typeOperator(
-                        resultType = unitType,
-                        argument = transformCall,
-                        typeOperator = IrTypeOperator.IMPLICIT_COERCION_TO_UNIT,
-                        typeOperand = unitType,
-                    )
-                }
+    private fun IrSimpleFunction.createReturnUnitBody() =
+        DeclarationIrBuilder(symbol).irBlockBody {
+            val unitClass: IrClassSymbol = context.irBuiltIns.unitClass.owner.symbol
+            val unitCall: IrGetObjectValue = irGetObject(unitClass)
+            +irReturn(unitCall)
         }
-        return function
+
+    context(context: IrPluginContext)
+    private fun transformUpdateFunction(declaration: IrSimpleFunction): IrSimpleFunction {
+        declaration.body = declaration.createUpdateBody()
+        return declaration
+    }
+
+    context(context: IrPluginContext)
+    private fun IrSimpleFunction.createUpdateBody() =
+        DeclarationIrBuilder(symbol).irBlockBody {
+            val transformCall: IrMemberAccessExpression<*> =
+                createTransformCall(this@createUpdateBody)
+            +coerceToUnit(transformCall)
+        }
+
+    private fun IrBlockBodyBuilder.createTransformCall(
+        declaration: IrSimpleFunction,
+    ): IrMemberAccessExpression<*> {
+        val thisGet: IrGetValue = irGet(declaration.extensionReceiver!!)
+        val transformValueParameter: IrValueParameter = declaration.regularParameters.first()
+        val invokeFunction: IrSimpleFunction = function1InvokeFunction()
+        val transformCallDispatchReceiver: IrGetValue =
+            irGetAsVariableFunction(transformValueParameter)
+
+        return irCallOp(
+            callee = invokeFunction.symbol,
+            type = declaration.typeParameters.first().defaultType,
+            dispatchReceiver = transformCallDispatchReceiver,
+            argument = thisGet,
+            origin = IrStatementOrigin.INVOKE,
+        )
     }
 
     context(context: IrPluginContext)
     private fun transformUpdateEachFunction(declaration: IrSimpleFunction): IrSimpleFunction {
-        val function: IrSimpleFunction = declaration.apply {
-            body =
-                DeclarationIrBuilder(declaration.symbol).irBlockBody {
-                    val thisGet: IrGetValue = irGet(extensionReceiver!!)
-                    val transformValueParameter: IrValueParameter = regularParameters.first()
-                    val transformParameterIndex: Int = transformValueParameter.indexInParameters - 1
-
-                    val transformCallType: IrSimpleType =
-                        listClass.typeWith(typeParameters.map { it.defaultType })
-
-                    val typeParameterType: IrSimpleType =
-                        declaration.typeParameters.map { it.defaultType }.first()
-
-                    val transformCall: IrCall =
-                        irCall(callee = mapFunction.symbol).apply {
-                            typeArguments[0] = typeParameterType
-                            typeArguments[1] = typeParameterType
-                            insertExtensionReceiver(thisGet)
-                            arguments[transformParameterIndex] = irGet(transformValueParameter)
-                            this.type = transformCallType
-                        }
-
-                    +typeOperator(
-                        resultType = unitType,
-                        argument = transformCall,
-                        typeOperator = IrTypeOperator.IMPLICIT_COERCION_TO_UNIT,
-                        typeOperand = unitType,
-                    )
-                }
-        }
-        return function
+        declaration.body = declaration.createUpdateEachBody()
+        return declaration
     }
 
-    private fun IrBlockBodyBuilder.thisCopyIrVariable(declaration: IrSimpleFunction): IrVariable {
-        val copyFun: IrSimpleFunction =
-            declaration.parentAsClass.findDeclaration<IrSimpleFunction> {
-                val isKopyCopyFun: Boolean = it.isKopyCopy
-                it.name == copyName && !isKopyCopyFun
-            }!!
+    context(context: IrPluginContext)
+    private fun IrSimpleFunction.createUpdateEachBody() =
+        DeclarationIrBuilder(symbol).irBlockBody {
+            val mapCall: IrCall = createMapCall(this@createUpdateEachBody)
+            +coerceToUnit(mapCall)
+        }
+
+    private fun IrBlockBodyBuilder.createMapCall(declaration: IrSimpleFunction): IrCall {
+        val thisGet: IrGetValue = irGet(declaration.extensionReceiver!!)
+        val transformValueParameter: IrValueParameter = declaration.regularParameters.first()
+        val typeParameterType: IrSimpleType = declaration.firstDefaultTypeParameter()
+
+        return irCall(callee = mapFunction.symbol).apply {
+            typeArguments[0] = typeParameterType
+            typeArguments[1] = typeParameterType
+            insertExtensionReceiver(thisGet)
+            arguments[transformValueParameter.indexInParameters - 1] =
+                irGet(transformValueParameter)
+            type = declaration.listOfDefaultTypeParameters()
+        }
+    }
+
+    private fun IrBlockBodyBuilder.coerceToUnit(expression: IrMemberAccessExpression<*>) =
+        typeOperator(
+            resultType = unitType,
+            argument = expression,
+            typeOperator = IrTypeOperator.IMPLICIT_COERCION_TO_UNIT,
+            typeOperand = unitType,
+        )
+
+    private fun IrBlockBodyBuilder.createTemporaryCopyVariable(
+        declaration: IrSimpleFunction,
+    ): IrVariable {
+        val copyFun: IrSimpleFunction = declaration.parentAsClass.kotlinDataClassCopyFunction()
         val receiver: IrGetValue = irGet(declaration.dispatchReceiverParameter!!)
         val copyCall: IrMemberAccessExpression<*> =
             irCall(copyFun.symbol).apply { dispatchReceiver = receiver }
@@ -197,13 +201,17 @@ internal class IrFunctionsTransformer(
         return irVariable
     }
 
-    private fun IrBlockBodyBuilder.copyCall(
+    private fun IrClass.kotlinDataClassCopyFunction(): IrSimpleFunction =
+        findDeclaration {
+            val isKopyCopyFun: Boolean = it.isKopyCopy
+            it.name == copyName && !isKopyCopyFun
+        }!!
+
+    private fun IrBlockBodyBuilder.callCopyLambda(
         declaration: IrSimpleFunction,
         thisCopyIrVariable: IrVariable,
     ): IrFunctionAccessExpression {
-        val invokeFunc: IrSimpleFunction =
-            function1Class.findDeclaration<IrSimpleFunction> { it.name == invokeName }!!
-
+        val invokeFunc: IrSimpleFunction = function1InvokeFunction()
         val copyValueParameter: IrValueParameter = declaration.regularParameters.first()
         val thisCopyIrVariableGetValue: IrGetValue = irGet(thisCopyIrVariable)
 
@@ -219,29 +227,42 @@ internal class IrFunctionsTransformer(
         return copyCall
     }
 
+    private fun IrBlockBodyBuilder.irGetAsVariableFunction(
+        valueParameter: IrValueParameter
+    ): IrGetValue = irGet(valueParameter).apply { origin = IrStatementOrigin.VARIABLE_AS_FUNCTION }
+
+    private fun function1InvokeFunction(): IrSimpleFunction =
+        function1Class.findDeclaration<IrSimpleFunction> { it.name == invokeName }!!
+
+    private fun IrSimpleFunction.firstDefaultTypeParameter(): IrSimpleType =
+        typeParameters.map { it.defaultType }.first()
+
+    private fun IrSimpleFunction.listOfDefaultTypeParameters(): IrSimpleType =
+        listClass.typeWith(typeParameters.map { it.defaultType })
+
     context(context: IrPluginContext)
-    private fun atomicValueIrReturn(
+    private fun returnAtomicValue(
         declaration: IrSimpleFunction,
         thisCopyIrVariable: IrVariable,
     ): IrReturn =
         DeclarationIrBuilder(declaration.symbol).run {
             val thisCopyGetValue: IrGetValue = irGet(thisCopyIrVariable)
-            val atomicProperty: IrProperty? =
-                declaration.parent.asIr<IrClass>().findDeclaration<IrProperty> {
-                    it.name == underscoreAtomicName
-                }
-
-            val getAtomicCall =
-                irCall(atomicProperty!!.getter!!).apply { dispatchReceiver = thisCopyGetValue }
-
-            val atomicReferenceClass: IrClass = firstIrClass(atomicReferenceClassId)
-
-            val atomicReferenceLoadFunction: IrFunction =
-                atomicReferenceClass.findDeclaration<IrFunction> { it.name == loadName }!!
-
-            val atomicReferenceLoadCall =
+            val atomicProperty: IrProperty = declaration.parent.asIr<IrClass>().atomicProperty()
+            val getAtomicCall: IrFunctionAccessExpression =
+                irCall(atomicProperty.getter!!).apply { dispatchReceiver = thisCopyGetValue }
+            val atomicReferenceLoadFunction: IrFunction = atomicReferenceLoadFunction()
+            val atomicReferenceLoadCall: IrFunctionAccessExpression =
                 irCall(atomicReferenceLoadFunction).apply { dispatchReceiver = getAtomicCall }
-            val irReturn: IrReturn = irReturn(atomicReferenceLoadCall)
-            irReturn
+            irReturn(atomicReferenceLoadCall)
         }
+
+    context(context: IrPluginContext)
+    private fun atomicReferenceLoadFunction(): IrFunction =
+        firstIrClass(atomicReferenceClassId).atomicReferenceLoadFunction()
+
+    private fun IrClass.atomicProperty(): IrProperty =
+        findDeclaration<IrProperty> { it.name == underscoreAtomicName }!!
+
+    private fun IrClass.atomicReferenceLoadFunction(): IrFunction =
+        findDeclaration<IrFunction> { it.name == loadName }!!
 }