+++ kopy-compiler/test-gen/java/com/javiersc/kotlin/kopy/compiler/BoxTestGenerated.java
@@ -16,6 +16,10 @@ import java.util.regex.Pattern;
 @TestMetadata("test-data/box")
 @TestDataPath("$PROJECT_ROOT")
 public class BoxTestGenerated extends AbstractBoxTest {
+  private void run(String fileName) {
+    runTest("test-data/box/" + fileName);
+  }
+
   @Test
   public void testAllFilesPresentInBox() {
     KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("test-data/box"), Pattern.compile("^(.+)\\.kt$"), null, true);
@@ -24,61 +28,65 @@ public class BoxTestGenerated extends AbstractBoxTest {
   @Test
   @TestMetadata("nest-copy-set.kt")
   public void testNest_copy_set() {
-    runTest("test-data/box/nest-copy-set.kt");
+    run("nest-copy-set.kt");
   }
 
   @Test
   @TestMetadata("nest-copy-update.kt")
   public void testNest_copy_update() {
-    runTest("test-data/box/nest-copy-update.kt");
+    run("nest-copy-update.kt");
   }
 
   @Test
   @TestMetadata("no-nest-copy-set.kt")
   public void testNo_nest_copy_set() {
-    runTest("test-data/box/no-nest-copy-set.kt");
+    run("no-nest-copy-set.kt");
   }
 
   @Test
   @TestMetadata("no-nest-copy-update-each-no-it.kt")
   public void testNo_nest_copy_update_each_no_it() {
-    runTest("test-data/box/no-nest-copy-update-each-no-it.kt");
+    run("no-nest-copy-update-each-no-it.kt");
   }
 
   @Test
   @TestMetadata("no-nest-copy-update-each-with-custom-parameter.kt")
   public void testNo_nest_copy_update_each_with_custom_parameter() {
-    runTest("test-data/box/no-nest-copy-update-each-with-custom-parameter.kt");
+    run("no-nest-copy-update-each-with-custom-parameter.kt");
   }
 
   @Test
   @TestMetadata("no-nest-copy-update-each-with-it.kt")
   public void testNo_nest_copy_update_each_with_it() {
-    runTest("test-data/box/no-nest-copy-update-each-with-it.kt");
+    run("no-nest-copy-update-each-with-it.kt");
   }
 
   @Test
   @TestMetadata("no-nest-copy-update-no-it.kt")
   public void testNo_nest_copy_update_no_it() {
-    runTest("test-data/box/no-nest-copy-update-no-it.kt");
+    run("no-nest-copy-update-no-it.kt");
   }
 
   @Test
   @TestMetadata("no-nest-copy-update-with-custom-parameter.kt")
   public void testNo_nest_copy_update_with_custom_parameter() {
-    runTest("test-data/box/no-nest-copy-update-with-custom-parameter.kt");
+    run("no-nest-copy-update-with-custom-parameter.kt");
   }
 
   @Test
   @TestMetadata("no-nest-copy-update-with-it.kt")
   public void testNo_nest_copy_update_with_it() {
-    runTest("test-data/box/no-nest-copy-update-with-it.kt");
+    run("no-nest-copy-update-with-it.kt");
   }
 
   @Nested
   @TestMetadata("test-data/box/edge")
   @TestDataPath("$PROJECT_ROOT")
   public class Edge {
+    private void run(String fileName) {
+      runTest("test-data/box/edge/" + fileName);
+    }
+
     @Test
     public void testAllFilesPresentInEdge() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("test-data/box/edge"), Pattern.compile("^(.+)\\.kt$"), null, true);
@@ -87,13 +95,13 @@ public class BoxTestGenerated extends AbstractBoxTest {
     @Test
     @TestMetadata("no-kopy-annotation-on-one-data-class.kt")
     public void testNo_kopy_annotation_on_one_data_class() {
-      runTest("test-data/box/edge/no-kopy-annotation-on-one-data-class.kt");
+      run("no-kopy-annotation-on-one-data-class.kt");
     }
 
     @Test
     @TestMetadata("simple-1.kt")
     public void testSimple_1() {
-      runTest("test-data/box/edge/simple-1.kt");
+      run("simple-1.kt");
     }
   }
 
@@ -101,6 +109,10 @@ public class BoxTestGenerated extends AbstractBoxTest {
   @TestMetadata("test-data/box/inheritance")
   @TestDataPath("$PROJECT_ROOT")
   public class Inheritance {
+    private void run(String fileName) {
+      runTest("test-data/box/inheritance/" + fileName);
+    }
+
     @Test
     public void testAllFilesPresentInInheritance() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("test-data/box/inheritance"), Pattern.compile("^(.+)\\.kt$"), null, true);
@@ -109,13 +121,13 @@ public class BoxTestGenerated extends AbstractBoxTest {
     @Test
     @TestMetadata("interface-1.kt")
     public void testInterface_1() {
-      runTest("test-data/box/inheritance/interface-1.kt");
+      run("interface-1.kt");
     }
 
     @Test
     @TestMetadata("sealed-class-1.kt")
     public void testSealed_class_1() {
-      runTest("test-data/box/inheritance/sealed-class-1.kt");
+      run("sealed-class-1.kt");
     }
   }
 
@@ -123,6 +135,10 @@ public class BoxTestGenerated extends AbstractBoxTest {
   @TestMetadata("test-data/box/nested-copy")
   @TestDataPath("$PROJECT_ROOT")
   public class Nested_copy {
+    private void run(String fileName) {
+      runTest("test-data/box/nested-copy/" + fileName);
+    }
+
     @Test
     public void testAllFilesPresentInNested_copy() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("test-data/box/nested-copy"), Pattern.compile("^(.+)\\.kt$"), null, true);
@@ -131,7 +147,7 @@ public class BoxTestGenerated extends AbstractBoxTest {
     @Test
     @TestMetadata("simple-1.kt")
     public void testSimple_1() {
-      runTest("test-data/box/nested-copy/simple-1.kt");
+      run("simple-1.kt");
     }
   }
 
@@ -139,6 +155,10 @@ public class BoxTestGenerated extends AbstractBoxTest {
   @TestMetadata("test-data/box/repeated-properties")
   @TestDataPath("$PROJECT_ROOT")
   public class Repeated_properties {
+    private void run(String fileName) {
+      runTest("test-data/box/repeated-properties/" + fileName);
+    }
+
     @Test
     public void testAllFilesPresentInRepeated_properties() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("test-data/box/repeated-properties"), Pattern.compile("^(.+)\\.kt$"), null, true);
@@ -147,13 +167,13 @@ public class BoxTestGenerated extends AbstractBoxTest {
     @Test
     @TestMetadata("complex-1.kt")
     public void testComplex_1() {
-      runTest("test-data/box/repeated-properties/complex-1.kt");
+      run("complex-1.kt");
     }
 
     @Test
     @TestMetadata("complex-2.kt")
     public void testComplex_2() {
-      runTest("test-data/box/repeated-properties/complex-2.kt");
+      run("complex-2.kt");
     }
   }
 
@@ -161,6 +181,10 @@ public class BoxTestGenerated extends AbstractBoxTest {
   @TestMetadata("test-data/box/serialization")
   @TestDataPath("$PROJECT_ROOT")
   public class Serialization {
+    private void run(String fileName) {
+      runTest("test-data/box/serialization/" + fileName);
+    }
+
     @Test
     public void testAllFilesPresentInSerialization() {
       KtTestUtil.assertAllTestsPresentByMetadataWithExcluded(this.getClass(), new File("test-data/box/serialization"), Pattern.compile("^(.+)\\.kt$"), null, true);
@@ -169,7 +193,7 @@ public class BoxTestGenerated extends AbstractBoxTest {
     @Test
     @TestMetadata("simple-serialization.kt")
     public void testSimple_serialization() {
-      runTest("test-data/box/serialization/simple-serialization.kt");
+      run("simple-serialization.kt");
     }
   }
 }